local/rest_version/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
matplotlib
pillow
numpy
opencv-python
torch
torchvision
rembg
rich
onnxruntime
openpyxl